import de controllers e main.py para controlar GUI
