+++ ProjetoLojaOnline_v02/src/controllers/user_controller.py
@@ -1,8 +1,8 @@
 # Caio Rabinovich Panes Brunholi
 # RA: 20.01285-3
 from hashlib import new
-from models.user import User
-from models.credit_card import CreditCard
+from src.models.user import User
+from src.models.credit_card import CreditCard
 
 # Login usando o user_controller ainda não implementado
