+++ ProjetoLojaOnline_v02/src/main.py
@@ -1,0 +1,5 @@
+from src.views.login import show_login_page
+import streamlit as st
+
+show_login_page()
+
